Remove db folder and refactor historian.py in MongodbHistorian. Remove Mongo from SQLHistorian.
