--- services/core/SQLHistorian/config.mongodb.platform.historian
@@ -1,14 +1,0 @@
-{
-    "agentid": "sqlhistorian-mongodb",
-    "identity": "platform.historian",
-    "connection": {
-        "type": "mongodb",
-        "params": {
-            "host": "10.211.55.2",
-            "port": 27017,
-            "database": "volttron_guide",
-            "user": "admin",
-            "passwd": "admin"
-        }
-    }
-}
